third_party/WebKit/LayoutTests/fast/css/getComputedStyle/computed-style-non-element-parameter.html

<script src="../../../resources/js-test.js"></script>
<script>
shouldThrow("getComputedStyle(null)");
shouldThrow("getComputedStyle(undefined)");
shouldThrow("getComputedStyle(document.createTextNode(\"foo\"))");
</script>
