third_party/blink/web_tests/platform/generic/fast/css/getComputedStyle/computed-style-non-element-parameter-expected.txt

PASS getComputedStyle(null) threw exception TypeError: Failed to execute 'getComputedStyle' on 'Window': parameter 1 is not of type 'Element'..
PASS getComputedStyle(undefined) threw exception TypeError: Failed to execute 'getComputedStyle' on 'Window': parameter 1 is not of type 'Element'..
PASS getComputedStyle(document.createTextNode("foo")) threw exception TypeError: Failed to execute 'getComputedStyle' on 'Window': parameter 1 is not of type 'Element'..
PASS successfullyParsed is true

TEST COMPLETE
